각종 오류 수정중
- 각종 참조 오류 계속 발생중

+++ Assets/My_LGS/UI/Scripts/PlayerData.cs
@@ -5,10 +5,10 @@ public class PlayerData
 {
     public Vector3 playerPosition;  // 플레이어의 위치
     public Vector3 playerrotation;  // 플레이어의 회전
-    public float playerHP;          // 플레이어의 체력
+    public int playerHP;          // 플레이어의 체력
     public int bulletCnt;           // 플레이어의 탄약 수
 
-    public void SaveData(Transform playerTransform, float hp, int bullets)
+    public void SaveData(Transform playerTransform, int hp, int bullets)
     {
         // 데이터 세이브
         playerPosition = playerTransform.position; // 플레이어의 위치 저장
@@ -17,14 +17,14 @@ public class PlayerData
         bulletCnt = bullets; // 플레이어의 탄약 수 저장
     }
 
-    public void SaveHPandBullet(float hp, int bullets)
+    public void SaveHPandBullet(int hp, int bullets)
     {
         // 데이터 세이브
         playerHP = hp; // 플레이어의 체력 저장
         bulletCnt = bullets; // 플레이어의 탄약 수 저장
     }
 
-    public void LoadData(Transform playerTransform, ref float hp, ref int bullets)
+    public void LoadData(Transform playerTransform, ref int hp, ref int bullets)
     {
         //데이터 로드
         playerTransform.position = playerPosition; // 저장된 위치로 플레이어 이동
@@ -33,7 +33,7 @@ public class PlayerData
         bullets = bulletCnt; // 저장된 탄약 수
     }
 
-    public void LoadHPandBullet(ref float hp, ref int bullets)
+    public void LoadHPandBullet(ref int hp, ref int bullets)
     {
         //데이터 로드
         hp = playerHP; // 저장된 체력
